pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "local-llm-manager"
version = "0.1.0"
description = "Intelligent hardware-aware tool for managing local LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Arjun Arihant" }
]
keywords = ["llm", "ollama", "local-ai", "gpu", "benchmark"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "GPUtil>=1.4.0",
    "psutil>=5.9.0",
    "py-cpuinfo>=9.0.0",
]

[project.scripts]
llm-manager = "local_llm_manager.cli:main"

[project.urls]
Homepage = "https://github.com/arjun-arihant/local-llm-manager"
Repository = "https://github.com/arjun-arihant/local-llm-manager"

[tool.hatch.build.targets.wheel]
packages = ["src/local_llm_manager"]

[tool.hatch.metadata]
allow-direct-references = true
